// interfaces/boat.interface.ts
export interface Boats {
    boats: Boat[];
}

export interface Boat {
    id_boat:       number;
    type:          string;
    model:         string;
    brand:         string;
    year:          number;
    length:        string;
    boat_position: string;
    person_name:   string;
    lastname:      string;
}